token取得エラーのアラート表示 判定処理の共通化対応
